+++ CHANGELOG.md
@@ -23,9 +23,10 @@ Per Keep a Changelog there are 6 main categories of changes:
 
 - Added a `key_value_data` function to the reader that returns an iterator over key-value pairs (by @expenses).
 - `Reader::levels` now returns an iterator over `Level` structs, which contain the bytes of the level as well as the uncompressed length (by @expenses).
-- Added `Header::from_bytes` and `LevelIndex::from_bytes` (by @expenses).
+- Added `Header::from_bytes`, `Header::as_bytes`, `LevelIndex::from_bytes` and `LevelIndex::as_bytes` (by @expenses).
 - Made the following fields public (by @expenses):
   - `Header::LENGTH`
+  - `Header::index`
   - `LevelIndex::LENGTH`
   - `LevelIndex::byte_offset`
   - `LevelIndex::byte_length`